conda_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 24.11.0
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
_sysroot_linux-64_curr_repodata_hack=3=haa98f57_10
accelerate=1.2.0=pyhd8ed1ab_0
aiohappyeyeballs=2.4.4=pyhd8ed1ab_1
aiohttp=3.11.10=pypi_0
aiosignal=1.3.1=pyhd8ed1ab_1
attrs=24.2.0=pyh71513ae_1
aws-c-auth=0.7.22=h96bc93b_2
aws-c-cal=0.6.14=h88a6e22_1
aws-c-common=0.9.19=h4ab18f5_0
aws-c-compression=0.2.18=h83b837d_6
aws-c-event-stream=0.4.2=ha47c788_12
aws-c-http=0.8.1=h29d6fba_17
aws-c-io=0.14.8=h21d4f22_5
aws-c-mqtt=0.10.4=h759edc4_4
aws-c-s3=0.5.9=h594631b_3
aws-c-sdkutils=0.1.16=h83b837d_2
aws-checksums=0.1.18=h83b837d_6
aws-crt-cpp=0.26.9=he3a8b3b_0
aws-sdk-cpp=1.11.329=hba8bd5f_3
binutils_impl_linux-64=2.40=h5293946_0
binutils_linux-64=2.40.0=hc2dff05_1
bitsandbytes=0.45.0=cuda118_py311h2c6edaf_0
blas=1.0=mkl
bottleneck=1.4.2=py311hf4808d0_0
brotli-python=1.1.0=py311hfdbb021_2
bzip2=1.0.8=h5eee18b_6
c-ares=1.34.3=hb9d3cd8_1
ca-certificates=2024.11.26=h06a4308_0
certifi=2024.8.30=pyhd8ed1ab_0
cffi=1.17.1=py311hf29c0ef_0
charset-normalizer=3.4.0=pyhd8ed1ab_1
colorama=0.4.6=pyhd8ed1ab_1
cudatoolkit=11.8.0=h6a678d5_0
cudnn=8.9.2.26=cuda11_0
cupti=11.8.0=he078b1a_0
dashscope=1.20.14=pypi_0
datasets=3.1.0=pyhff2d567_0
dill=0.3.8=pyhd8ed1ab_0
filelock=3.13.1=py311h06a4308_0
frozenlist=1.5.0=py311h9ecbd09_0
fsspec=2024.6.1=py311h06a4308_0
gcc_impl_linux-64=11.2.0=h1234567_1
gcc_linux-64=11.2.0=h5c386dc_1
gflags=2.2.2=h5888daf_1005
glog=0.7.1=hbabe93e_0
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py311hc9b5ff0_0
gxx_impl_linux-64=11.2.0=h1234567_1
gxx_linux-64=11.2.0=hc2dff05_1
h2=4.1.0=pyhd8ed1ab_1
hpack=4.0.0=pyhd8ed1ab_1
huggingface_hub=0.26.5=pyhd8ed1ab_0
hyperframe=6.0.1=pyhd8ed1ab_1
idna=3.10=pyhd8ed1ab_1
intel-openmp=2023.1.0=hdb19cb5_46306
jinja2=3.1.4=py311h06a4308_1
joblib=1.4.2=pypi_0
kernel-headers_linux-64=3.10.0=h57e8cba_10
keyutils=1.6.1=h166bdaf_0
krb5=1.21.3=h659f571_0
ld_impl_linux-64=2.40=h12ee557_0
libabseil=20240116.2=cxx17_h6a678d5_0
libarrow=16.1.0=hcb6531f_6_cpu
libarrow-acero=16.1.0=hac33072_6_cpu
libarrow-dataset=16.1.0=hac33072_6_cpu
libarrow-substrait=16.1.0=h7e0c224_6_cpu
libbrotlicommon=1.1.0=hb9d3cd8_2
libbrotlidec=1.1.0=hb9d3cd8_2
libbrotlienc=1.1.0=hb9d3cd8_2
libcrc32c=1.1.2=h9c3ff4c_0
libcurl=8.8.0=hca28451_1
libedit=3.1.20191231=he28a2e2_2
libev=4.33=hd590300_2
libevent=2.1.12=hf998b51_1
libffi=3.4.4=h6a678d5_1
libgcc=14.2.0=h77fa898_1
libgcc-devel_linux-64=11.2.0=h1234567_1
libgcc-ng=14.2.0=h69a702a_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=14.2.0=h77fa898_1
libgoogle-cloud=2.24.0=h2736e30_0
libgoogle-cloud-storage=2.24.0=h3d9a0c8_0
libgrpc=1.62.2=h15f2491_0
libnghttp2=1.58.0=h47da74e_1
libparquet=16.1.0=h6a7eafb_6_cpu
libprotobuf=4.25.3=he621ea3_0
libre2-11=2023.09.01=h5a48ba9_2
libssh2=1.11.0=h0841786_0
libstdcxx=14.2.0=hc0a3c3a_1
libstdcxx-devel_linux-64=11.2.0=h1234567_1
libstdcxx-ng=14.2.0=h4852527_1
libthrift=0.19.0=hb90f79a_1
libutf8proc=2.8.0=hf23e847_1
libuuid=1.41.5=h5eee18b_0
libzlib=1.2.13=h4ab18f5_6
lz4-c=1.9.4=hcb278e6_0
magma=2.7.1=h2c23e93_0
markupsafe=2.1.3=py311h5eee18b_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py311h5eee18b_1
mkl_fft=1.3.11=py311h5eee18b_0
mkl_random=1.2.8=py311ha02d727_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=py311h06a4308_0
multidict=6.1.0=py311h2dc5d0c_1
multiprocess=0.70.16=py311h9ecbd09_1
ncurses=6.4=h6a678d5_0
networkx=3.2.1=py311h06a4308_0
numexpr=2.10.1=py311h3c60e43_0
numpy=2.0.1=py311h08b1b3b_1
numpy-base=2.0.1=py311hf175353_1
openssl=3.4.0=hb9d3cd8_0
orc=2.0.1=h17fec99_1
packaging=24.2=pyhd8ed1ab_2
pandas=2.2.3=py311h6a678d5_0
pillow=11.0.0=pypi_0
pip=24.2=py311h06a4308_0
propcache=0.2.1=py311h9ecbd09_0
psutil=5.9.0=py311h5eee18b_0
pyarrow=16.1.0=py311hbd00459_2
pyarrow-core=16.1.0=py311hf1d6e26_2_cpu
pycparser=2.22=pyh29332c3_1
pysocks=1.7.1=pyha55dd90_7
python=3.11.10=he870216_0
python-dateutil=2.9.0post0=py311h06a4308_2
python-tzdata=2023.3=pyhd3eb1b0_0
python-xxhash=3.5.0=py311h9ecbd09_1
python_abi=3.11=2_cp311
pytorch=2.3.1=gpu_cuda118py311hec12777_100
pytz=2024.1=py311h06a4308_0
pyyaml=6.0.2=py311h9ecbd09_1
re2=2023.09.01=h7f4b329_2
readline=8.2=h5eee18b_0
regex=2024.11.6=py311h9ecbd09_0
requests=2.32.3=pyhd8ed1ab_1
s2n=1.4.15=he19d79f_0
safetensors=0.4.5=py311h9e33e62_0
scikit-learn=1.6.0=pypi_0
scipy=1.14.1=py311h08b1b3b_0
sentence-transformers=3.3.1=pypi_0
setuptools=75.1.0=py311h06a4308_0
six=1.16.0=pyhd3eb1b0_1
snappy=1.2.1=h8bd8927_1
sqlite=3.45.3=h5eee18b_0
sympy=1.13.2=py311h06a4308_0
sysroot_linux-64=2.17=h57e8cba_10
tbb=2021.8.0=hdb19cb5_0
threadpoolctl=3.5.0=pypi_0
tk=8.6.14=h39e8969_0
tokenizers=0.20.4=py311h182c674_0
torchtriton=2.3.0=cuda123py311hdb19cb5_0
tqdm=4.67.1=pyhd8ed1ab_0
transformers=4.46.1=pyhd8ed1ab_0
typing-extensions=4.11.0=py311h06a4308_0
typing_extensions=4.11.0=py311h06a4308_0
tzdata=2024b=h04d1e81_0
urllib3=2.2.3=pyhd8ed1ab_1
websocket-client=1.8.0=pypi_0
wheel=0.44.0=py311h06a4308_0
xxhash=0.8.2=hd590300_0
xz=5.4.6=h5eee18b_1
yaml=0.2.5=h7f98852_2
yarl=1.18.3=py311h9ecbd09_0
zlib=1.2.13=h4ab18f5_6
zstandard=0.23.0=py311hbc35293_1
zstd=1.5.6=ha6fb4c9_0
